WIP (do not review) - CatBoost bridge
This commit moves the catboost model evaluation out of the server
process into a separate "bridge" binary. This serves two goals: On the
one hand, crashes / memory corruptions of the catboost library no longer
affect the server and 2. we can forbid loading dynamic libraries in the
server (catboost was the last consumer of this functionality), thus
improving security.

Bridges exchange data with the server via a RESTful API. Right now, two
bridge binaries exist (library-bridge and xdbc-bridge). I originally
considered extending the former with access to libcatboost but despite
it's generic name, it is tailored to external dictionaries. As a result,
thic commit introduces "clickhouse-catboost-brige" as third bridge.

The protocol is simple:
- HTTP GET: essentially a ping which returns if libcatboost is loaded,
  used during the server-bridge handshake to check if the bridge runs.
- HTTP POST
  - method "libNew": load libcatboost into bridge
  - method "libDelete": unload libcatboost from bridge
  - method "evaluate": send inputs (float and categorical features),
    then evaluate the model in bridge, then send results back to server

Stuff deleted in this commit:
- SQL function "modelEvaluate"
- monitoring view "SYSTEM.SYSTEMMODELS" (it wasn't documented)
- classes ExternalModelsLoader + IMLModel: this infrastructure managed
  shared libraries loaded at runtime and it's last raison d'être was
  catboost. The new SQL function "catboostEvaluate" is much more
  light-weight, it combines the load/evaluate/unload steps into one
  "atomic" operation.

TODO:
- update documentation
- integration tests